requirements.txt
django
python-nmap
pillow
netifaces
influxdb
ipaddress
